+++ Arc/src/Platform/OpenGL/OpenGLTexture.h
@@ -15,7 +15,7 @@ namespace ArcEngine
 		
 		virtual uint32_t GetWidth() const override { return m_Width; }
 		virtual uint32_t GetHeight() const override { return m_Height; }
-		virtual uint32_t GetRendererID() const override { return m_RendererID; }
+		virtual uint64_t GetRendererID() const override { return m_RendererID; }
 		virtual const eastl::string& GetPath() const override { return m_Path; }
 
 		virtual void SetData(void* data, uint32_t size) override;
