+++ Arc/src/Platform/OpenGL/OpenGLTextureCubemap.h
@@ -14,8 +14,8 @@ namespace ArcEngine
 		
 		virtual uint32_t GetWidth() const override { return m_Width; }
 		virtual uint32_t GetHeight() const override { return m_Height; }
-		virtual uint32_t GetRendererID() const override { return m_RendererID; }
-		virtual uint32_t GetHRDRendererID() const override { return m_HRDRendererID; }
+		virtual uint64_t GetRendererID() const override { return m_RendererID; }
+		virtual uint64_t GetHRDRendererID() const override { return m_HRDRendererID; }
 		virtual const eastl::string& GetPath() const override { return m_Path; }
 
 		virtual void SetData(void* data, uint32_t size) override;
